Reorganize loggers and imports.
