+++ clients/privacy-center/app/server-utils/logger.ts
@@ -1,0 +1,11 @@
+import pino from "pino";
+
+import { LogLevels } from "./PrivacyCenterSettings";
+
+export const createLogger = ({ logLevel }: { logLevel: LogLevels }) => {
+  const logger = pino({
+    level: logLevel,
+  });
+
+  return logger;
+};
