+++ clients/privacy-center/middleware.ts
@@ -1,10 +1,14 @@
 import { NextRequest, NextResponse } from "next/server";
 import { v4 } from "uuid";
 
-import { createLogger } from "./app/server-utils/loggerContext";
+import loadEnvironmentVariables from "./app/server-utils/loadEnvironmentVariables";
+import { createLogger } from "./app/server-utils/logger";
 
 export default function middleware(request: NextRequest) {
-  const log = createLogger();
+  const envVars = loadEnvironmentVariables();
+  const log = createLogger({
+    logLevel: envVars.LOG_LEVEL,
+  });
   const requestId = v4();
 
   const logDict = {
